change current_user, is_professor and is_student
change current user method to assume that session will contain a :professor_id or a :student_id rather than a :user_id,
and changed is_professor and is_student to simply check for the appropriate id in the session.
current_user should now return Professor or Student object depending on who's logged in.